requirements.txt
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
drf-yasg==1.21.5
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
six==1.16.0
whitenoise==6.3.0